+++ packages/field/src/Field.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/button-has-type */
 import * as React from 'react';
 import { classes, hasOwnProperty, isObject, noop, toKey } from '@swan-form/helpers';
-import asField from './asField';
+import asField, { InjectedProps } from './asField';
 
 export interface FieldProps {
   name: string;
@@ -9,14 +9,14 @@ export interface FieldProps {
 
   required?: boolean;
   id?: string;
-  errors: React.ReactNode[];
+  // errors?: React.ReactNode[];
   className?: string;
   label?: React.ReactNode;
   icon?: React.ReactNode;
-  setRef(el: any): void;
-  onChange(event: React.ChangeEvent<any>): void;
-  onBlur(event: React.FocusEvent<any>): void;
-  setValue(value: any): void;
+  // setRef?(el: any): void;
+  // onChange?(event: React.ChangeEvent<any>): void;
+  // onBlur?(event: React.FocusEvent<any>): void;
+  // setValue?(value: any): void;
   style?: React.CSSProperties;
   options?: any;
   /**
@@ -61,7 +61,7 @@ const emptyObject = {};
 /**
  * Gets the classes applied to the outermost node in the field
  */
-export const getFieldClasses = (props: Partial<FieldProps> & { isSet?: boolean }) => {
+export const getFieldClasses = (props: Partial<FieldProps & InjectedProps> & { isSet?: boolean }) => {
   const { type, errors, className, required, icon, isSet } = props;
   return classes(
     isSet ? 'sf--fieldset' : 'sf--field',
@@ -155,10 +155,8 @@ function renderOptions(options: any | any[]): React.ReactNode {
 }
 /* eslint-enable no-use-before-define */
 
-export class Field extends React.PureComponent<FieldProps, {}> {
+export class FieldRender extends React.PureComponent<FieldProps & InjectedProps, {}> {
   static displayName = 'Field';
-
-  shouldWrapInLabel = () => !NO_WRAP.includes(this.props.type);
 
   getId = () => this.props.id || this.props.name;
 
@@ -176,6 +174,7 @@ export class Field extends React.PureComponent<FieldProps, {}> {
       hasFormSubmitted,
       ...props
     } = this.props;
+
     const id = this.getId();
 
     /**
@@ -222,21 +221,25 @@ export class Field extends React.PureComponent<FieldProps, {}> {
           <input {...props} className={submitResetClasses} type={type} id={id} ref={setRef} />
         );
       default:
-        return INPUT_TYPES.includes(type) ? (
-          <input {...props} className={appliedClasses} type={type} id={id} ref={setRef} />
-        ) : (
-          <input {...props} className={appliedClasses} type="text" id={id} ref={setRef} />
+        return (
+          <input
+            {...props}
+            className={appliedClasses}
+            type={INPUT_TYPES.includes(type) ? type : 'text'}
+            id={id}
+            ref={setRef}
+          />
         );
     }
   };
 
   render() {
     const id = this.getId();
-    const { errors, label, icon, style = emptyObject } = this.props;
+    const { errors, label, type, icon, style = emptyObject } = this.props;
 
     const appliedClasses = getFieldClasses(this.props);
 
-    if (this.shouldWrapInLabel()) {
+    if (!NO_WRAP.includes(type)) {
       return (
         <label htmlFor={id} className={appliedClasses} style={style}>
           <span>
@@ -253,4 +256,7 @@ export class Field extends React.PureComponent<FieldProps, {}> {
   }
 }
 
-export default asField(Field);
+// @TODO fix typings
+export const Field = asField(FieldRender);
+// export const Field = asField<FieldProps>(FieldRender);
+export default Field;